do not follow HTTP redirects
When a health check is defined, we want to probe the target system, not any it redirects us to. If we have a redirect, we want to test for a 302 status code.

Behavior without this change is that the golang http request library automatically follows redirects, and returns the target response's HTTP status code.

Fixes #75